refactor(random): Random Number를 반환하는 Util 클래스를 도메인 클래스로 변경
