+++ src/main/java/racingcar/domain/random/RandomNumberGenerator.java
@@ -1,13 +1,13 @@
-package racingcar.domain.global;
+package racingcar.domain.random;
 
 import camp.nextstep.edu.missionutils.Randoms;
 
-public abstract class RandomNumberUtils {
+public class RandomNumberGenerator {
 
     private static final int RANDOM_NUMBER_MIN = 0;
     private static final int RANDOM_NUMBER_MAX = 9;
 
-    public static RandomNumber getNumber() {
+    public RandomNumber create() {
         int number = Randoms.pickNumberInRange(RANDOM_NUMBER_MIN, RANDOM_NUMBER_MAX);
         return new RandomNumber(number);
     }
